xfree/requirements.txt
ipywidgets==8.1.3
matplotlib==3.9.2
numpy==2.0.1
openai==1.41.0
python-dotenv==1.0.1
scikit_learn==1.5.1
scipy==1.14.0
webvtt_py==0.5.1
pandas
seaborn